spark/requirements.txt
pyspark==3.5.1
kafka-python==2.0.2
Flask==3.0.3
